create files for each view controller in BITabBarController
