+++ Binar/Binar/PresentationLayer/BI/NewsChallenge/BIAnotherNewsViewController.swift
@@ -1,0 +1,19 @@
+//
+//  BIAnotherNewsViewController.swift
+//  Binar
+//
+//  Created by Bagas on 07/04/22.
+//
+
+import UIKit
+
+final class BIAnotherNewsViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        navigationItem.title = "News+"
+        navigationController?.navigationBar.prefersLargeTitles = true
+        view.backgroundColor = .systemBackground
+    }
+    
+}
