+++ Binar/Binar/PresentationLayer/BI/NewsChallenge/BINewsAudioViewController.swift
@@ -1,0 +1,20 @@
+//
+//  BINewsAudioViewController.swift
+//  Binar
+//
+//  Created by Bagas on 07/04/22.
+//
+
+import UIKit
+
+final class BINewsAudioViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        navigationItem.title = "Audio"
+        navigationController?.navigationBar.prefersLargeTitles = true
+        view.backgroundColor = .systemBackground
+    }
+    
+
+}
