+++ Binar/Binar/PresentationLayer/BI/NewsChallenge/BINewsFollowingViewController.swift
@@ -1,0 +1,18 @@
+//
+//  BINewsFollowingViewController.swift
+//  Binar
+//
+//  Created by Bagas on 07/04/22.
+//
+
+import UIKit
+
+final class BINewsFollowingViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        navigationItem.title = "Following"
+        navigationController?.navigationBar.prefersLargeTitles = true
+        view.backgroundColor = .systemBackground
+    }
+}
